backend/requirements.txt
Flask
flask-cors
google-cloud-vision
python-dotenv
twilio
gunicorn
google-auth